+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Settings/queries.yml
@@ -10,7 +10,7 @@ services:
     Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\IsConnectionsNumberLimitReachedQuery:
         arguments:
             - '@database_connection'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\Connections\GetConnectionsNumberLimit'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Connections\Service\GetConnectionsNumberLimit'
 
     Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository:
         arguments:
