+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/services.yml
@@ -1,6 +1,3 @@
-parameters:
-    akeneo_connectivity.connection.max_limit: 50
-
 services:
     akeneo_connectivity.connection.service.client.create_client:
         class: Akeneo\Connectivity\Connection\Infrastructure\Client\Fos\CreateClient
@@ -88,9 +85,5 @@ services:
         class: Akeneo\Connectivity\Connection\Infrastructure\ContentSecurityPolicy\MarketplaceContentSecurityPolicy
         tags:
             - { name: akeneo.pim.security.content_security_policy_provider }
-
-    Akeneo\Connectivity\Connection\Infrastructure\Connections\GetConnectionsNumberLimit:
-        arguments:
-            - '%akeneo_connectivity.connection.max_limit%'
 
     Akeneo\Connectivity\Connection\Infrastructure\Service\RandomCodeGenerator: ~
