+++ src/Oro/Bundle/DataGridBundle/Extension/Action/ActionExtension.php
@@ -150,7 +150,7 @@ class ActionExtension extends AbstractExtension
 
     /**
      * Creates and configure action object
-     * Services are marked as scope: prototype
+     * Service are marked as scope: prototype
      *
      * @param ActionConfiguration $config
      *
